+++ packages/manager/apps/web-domains/src/common/enum/zone.enum.ts
@@ -41,6 +41,7 @@ export enum RecordTypesTargetWithTrailingDotEnum {
   CNAME = 'CNAME',
   DNAME = 'DNAME',
   MX = 'MX',
+  SRV = 'SRV',
 }
 
 export enum NaptrFlagEnum {